download the images for the given species - WIP
The images are downloaded into the "images" folder, organized by taxon_key (speciesKey/taxonKey)
(the GBIF identifier for a species) and by occurrence_key (each of the specimens for the selected species).

Issues: Runnning a test using Quercus robur and Taxus baccata. About half of the images are not downloaded.
(total of 14 occurences por Q. robur and 12 for T. baccata).