get uk native tree names (dictionary)
Source: https://www.woodlandtrust.org.uk/trees-woods-and-wildlife/british-trees/native-trees/
Keys of the dictionary are the binomial names.
Values are the vernaular names in UK. Vernacular names are  not needed for the data request in GBIF, but added now because it was not too much extra work and may be useful later.