get information about the data request and data downloaded
A funtion to open a text file that will be used in the data_request.py in order to get a summary of the request and downloads.
Why: analyze how the filters applied in the request affect the data downloaded.
This will help to choose a suitable dataset to work with.